requirements.txt
gradio==3.45.2
transformers==4.47.1
pillow==10.4.0
huggingface_hub==0.26.3
python-dotenv==1.0.1
torch==2.2.2
torchvision==0.17.2